% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brainspan_id}
\alias{brainspan_id}
\title{Brainspan ID information}
\format{A data frame with 57 rows (57 ID's) and 9 columns. The ethnicity codes
key is not available at the present time.
\describe{
  \item{Stage}{integer to denote which stage of development the post-mortem brain is in}
  \item{Braincode}{ID}
  \item{Age}{age in PCW (post-conception weeks) or in M (months) or Y (years) after birth}
  \item{Days}{age but converted into days}
  \item{Gender}{gender}
  \item{Hemisphere}{L or R to denote if only the left or right hemisphere of the post-mortem
  brain was collected. Otherwise, both hemispheres available}
  \item{pH}{pH of the tissue}
  \item{Ethn.}{code to denote ethnicity}
  \item{PMI}{post-mortem interval}
}}
\source{
The dataset is based on
"Kang, H. J., Kawasawa, Y. I., Cheng, F., Zhu, Y., Xu, X., Li, M., Sousa,
A. M., Pletikos, M., Meyer, K. A., and Sedmak, G. (2011). Spatio-temporal
transcriptome of the human brain. Nature, 478(7370):483–489." The raw data
(prior to our lab's preprocessing) is located at \url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE25219}.
}
\description{
Brainspan ID information
}
\author{
Kevin Lin \email{kevinl1@andrew.cmu.edu}
}
\keyword{data}
