% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clique_selection.R
\name{clique_selection}
\alias{clique_selection}
\title{Find largest soft clique in graph}
\usage{
clique_selection(g, num_pos = c(1:4), num_neg = c(0:3),
  threshold = 0.95)
}
\arguments{
\item{g}{\code{igraph} object}

\item{num_pos}{numeric vector}

\item{num_neg}{numeric vector}

\item{threshold}{numeric}
}
\value{
list of indices
}
\description{
For a given graph \code{g} (igraph object) and a threshold \code{threshold}
(number between 0 and 1), find the largest set of nodes such that
the number of edges among this set of nodes in \code{g} is at least
\code{threshold} percentage of a full clique.
}
\details{
This is returned as a list of indices, so if there are multiple largest sets
of the same size, the list will have multiple set of indices.
}
