% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp_profiling.R
\name{compute_all_numerator_bootstrap_tmp1}
\alias{compute_all_numerator_bootstrap_tmp1}
\title{Something}
\usage{
compute_all_numerator_bootstrap_tmp1(dat_list, noise_list, cov_list,
  remaining_idx)
}
\arguments{
\item{dat_list}{Something}

\item{noise_list}{Something}

\item{cov_list}{Something}

\item{remaining_idx}{Something}
}
\value{
Something
}
\description{
Something
}
