% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_functions.R
\name{compute_scale_free}
\alias{compute_scale_free}
\title{Compute the scale-free statistic}
\usage{
compute_scale_free(adj)
}
\arguments{
\item{adj}{adjacency matrix}
}
\value{
numeric
}
\description{
Compute the scale-free statistic
}
