% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_functions.R
\name{count_neighbors}
\alias{count_neighbors}
\title{Count neighbors of a graph}
\usage{
count_neighbors(g, idx, order = 1)
}
\arguments{
\item{g}{igraph object}

\item{idx}{vector of indicies}

\item{order}{what degree neighbor}
}
\value{
a vector of numerics
}
\description{
Count how many neighbors each vertex in \code{idx} has.
}
