% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepdown.R
\name{.compute_all_denom}
\alias{.compute_all_denom}
\title{Compute all of the denominators}
\usage{
.compute_all_denom(dat_list, cov_list = as.list(rep(NA,
  length(dat_list))))
}
\arguments{
\item{dat_list}{list of data matrices}

\item{cov_list}{list of covariance matrices}
}
\value{
a list of denominator vectors
}
\description{
Compute all of the denominators
}
