% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepdown.R
\name{.compute_all_test_stat}
\alias{.compute_all_test_stat}
\title{Compute the test statistic}
\usage{
.compute_all_test_stat(num_list, denom_list,
  combn_mat = utils::combn(length(num_list), 2), squared = T)
}
\arguments{
\item{num_list}{list of numerator vectors}

\item{denom_list}{list of denominator vectors}

\item{combn_mat}{matrix of pairs to test for}

\item{squared}{boolean on whether or not to square the test statistic}
}
\value{
vector of all the bootstrap statistics
}
\description{
Compute the test statistic
}
