% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cai_test.R
\name{.compute_bootSigma}
\alias{.compute_bootSigma}
\title{Compute the bootstrap empirical covariance (numerator)}
\usage{
.compute_bootSigma(mat, noise_vec, cov_mat = NA, idx = NA)
}
\arguments{
\item{mat}{data matrix}

\item{noise_vec}{vector of noise}

\item{cov_mat}{covariance matrix}

\item{idx}{vector of the indices of the lower triangle}
}
\value{
vector
}
\description{
Compute the bootstrap empirical covariance (numerator)
}
