% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodness_of_fit.R
\name{.partition_data}
\alias{.partition_data}
\title{Enumerate partitions}
\usage{
.partition_data(n, num_partitions)
}
\arguments{
\item{n}{positive integer}

\item{num_partitions}{number of partitions}
}
\value{
list of partitions, each a list of 2 vector of integers
}
\description{
Enumerate partitions
}
