% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmrf.R
\name{hmrf}
\alias{hmrf}
\title{Hidden markov random field}
\usage{
hmrf(pval, adj, seedindex, pthres = 0.05, iter = 100,
  verbose = FALSE, tol = 0.001)
}
\arguments{
\item{pval}{a vector of d p-values, between 0 and 1.}

\item{adj}{a dxd adjancey matrix (0,1 entries)}

\item{seedindex}{a (0,1) vector of length d, where 1 means the gene is in group 1.}

\item{pthres}{threshold for p-values to serve as an initialization}

\item{iter}{number of iterations}

\item{verbose}{boolean}

\item{tol}{tolerance for values that should be treated as zero in absolute value}
}
\value{
a list
}
\description{
Use a Ising model to update the pvalues according to the graphical structure.
There are two groups, indices in group 0 have mean 0. Indices in group 1 have
mean greater than 0.
}
