% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonparanormal.R
\name{nonparanormal_transformation}
\alias{nonparanormal_transformation}
\title{Nonparanormal transformation}
\usage{
nonparanormal_transformation(dat, density_list, mean_vec, sd_vec)
}
\arguments{
\item{dat}{\code{n} by \code{d} matrix}

\item{density_list}{list of \code{d} \code{density} objects}

\item{mean_vec}{numeric vector of length \code{d}}

\item{sd_vec}{positive numeric vector of length \code{d}}
}
\value{
\code{n} by \code{d} matrix
}
\description{
Nonparanormal transformation
}
