% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_test.R
\name{pairwise_test}
\alias{pairwise_test}
\title{Pairwise Test Function}
\usage{
pairwise_test(dat_list, dat_list2 = NA, trials = 100, quant = 1,
  cores = 1)
}
\arguments{
\item{dat_list}{list of matrices}

\item{dat_list2}{(optional) list of matrices}

\item{trials}{number of trials}

\item{quant}{quantile}

\item{cores}{number of cores}
}
\value{
matrix of p-values
}
\description{
Uses Cai's covariance test between all the elements in a list.
Only the lower-triangular values are filled in.
}
\details{
If \code{dat_list2}
is provided, then instead this function apply's Cai's test
between every element in \code{dat_list} and \code{dat_list2}. In
this case, the returned matrix has \code{dat_list} elements
along the columns, and \code{dat_list2} elements along the rows.
}
