% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen.R
\name{screen}
\alias{screen}
\title{Screening data}
\usage{
screen(dat, pv, pthres = 0.1, num_genes = min(100,
  ceiling(nrow(dat)/10)))
}
\arguments{
\item{dat}{the nxd matrix}

\item{pv}{vector of d pvalues}

\item{pthres}{numeric}

\item{num_genes}{numeric}
}
\value{
list of primary and secondary genes, each a vector of integers
between 1 and \code{ncol(dat)}
}
\description{
We use a screening method where genes with \code{pv} lower than
\code{pthres} are primary genes. \code{num_genes} are selected, where
the remaining secondary genes are selected based on having the having
(absolute) Pearson correlation with any of the primary genes.
}
