% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clique_selection.R
\name{select_clique}
\alias{select_clique}
\title{Select clique based on intersection}
\usage{
select_clique(lis, idx, adj)
}
\arguments{
\item{lis}{list of indices}

\item{idx}{vector of indices}

\item{adj}{adjacency matrix}
}
\value{
one element of \code{lis}
}
\description{
Select clique based on intersection
}
