% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_clustering.R
\name{spectral_selection}
\alias{spectral_selection}
\title{Find largest clustering based on spectral clustering}
\usage{
spectral_selection(g, K = 2, target_idx = NA)
}
\arguments{
\item{g}{\code{igraph} object}

\item{K}{number of clusters}

\item{target_idx}{which indices to select cluster based on}
}
\value{
vector
}
\description{
Wrapper for \code{.spectral_cluster}
}
