% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming.R
\name{symbol_synonyms}
\alias{symbol_synonyms}
\title{Gene synonyms}
\usage{
symbol_synonyms(vec)
}
\arguments{
\item{vec}{vector of gene names (characters)}
}
\value{
vector of characters
}
\description{
Gene synonyms
}
